requirements.txt
flask>=2.3.0
pymongo>=4.6.0
dnspython>=2.4.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
reportlab>=4.0.0
python-dotenv>=1.0.0
phonenumbers>=8.13.0
